working on db setup
